tests/mediacapture-depth/BAK/MediaStreamTrack-attributes-manual.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Test MediaStreamTrack attributes</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://w3c.github.io/mediacapture-main/#dom-mediastream-getvideotracks">
<link rel="help" href="https://w3c.github.io/mediacapture-main/#dom-mediastreamtrack-label">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<meta name="flags" content="interact">

<h2>Preconditions</h2>
<ol>
  <li>
    Use a test device with depth camera(embedded or external) and an audio device.
  </li>
  <li>
    When prompted, accept to share your audio and depth streams.
  </li>
</ol>

<div id="log"></div>

<script>

promise_test(t => {
  let video;
  return navigator.mediaDevices.getUserMedia({video: {videoKind: {exact: "depth"}}})
  .then(stream => {
    video = stream;
    return navigator.mediaDevices.getUserMedia({audio: true})
    .then(audio => {
      video.addTrack(audio.getAudioTracks()[0]);
      assert_not_equals(video.getVideoTracks()[0].id, video.getAudioTracks()[0].id, "audio and video tracks have distinct ids");
    });
  });
}, "Tests that distinct mediastream tracks have distinct ids");

promise_test(t => {
  let id, label, kind;
  return navigator.mediaDevices.getUserMedia({video: {videoKind: {exact: "depth"}}})
  .then(stream => {
    track = stream.getTracks()[0];
    id = track.getSettings().deviceId;
    label = track.label;
    kind = track.kind;
    return navigator.mediaDevices.enumerateDevices();
  })
  .then(devices => {
    let device_label, device_kind;
    for (let i = 0; i < devices.length; ++i) {
      if (devices[i].deviceId == id) {
        device_label = devices[i].label;
        device_kind = devices[i].kind;
        break;
      }
    }
    assert_equals(device_label, label);
    assert_equals(kind, "video");
    assert_equals(device_kind, "videoinput");
  });
}, "Tests that the label and kind attributes must return the values of the object's corresonding source");

</script>
